Switch isConnected fn to middleware
